moved readcountrydb typing to utils and fixed imports
